refactor: divide by components
